Print warning when table does not exist
In some cases (such as when running rake tasks to create a database),
the potential_crufts table may not exist. In these cases, is_this_used
should print a warning and not blow things up. This commit updates the
`is_this_used?` method to catch this error and print the warning.